+++ conanfile.py
@@ -6,11 +6,6 @@ import re
 import os
 
 
-def to_camel_case(input: str):
-    words = input.replace("_", " ").split()
-    return '_'.join(word.capitalize() for word in words)
-
-
 class PackageConan(ConanFile):
     # @+ START USER META CONFIG
     license = "Apache 2.0"
@@ -19,10 +14,16 @@ class PackageConan(ConanFile):
     description = ""
     topics = ("")
     settings = "os", "compiler", "build_type", "arch"
-    options = {}
-    default_options = {}
+    options = {"shared": [True, False],
+               "fPIC": [True, False]}
+    default_options = {"shared": True,
+                       "fPIC": True}
     default_user = "Hahn-Schickard"
     # @- END USER META CONFIG
+    exports = [
+        "CMakeLists.txt",
+        "conanfile.py"
+    ]
     exports_sources = [
         "cmake*",
         "includes*",
@@ -34,17 +35,21 @@ class PackageConan(ConanFile):
         # @- END USER EXPORTS
     ]
     generators = "CMakeDeps"
+    package_type = "library"
     short_paths = True
-    package_type = "header-library"
 
     @property
     def cwd(self):
         return os.path.dirname(os.path.realpath(__file__))
 
+    @property
+    def full_name(self):
+        content = load(self, path=os.path.join(
+            self.recipe_folder, 'CMakeLists.txt'))
+        return re.search('set\(THIS (.*)\)', content).group(1).strip()
+
     def set_name(self):
-        content = load(self, path=os.path.join(self.cwd, 'CMakeLists.txt'))
-        name = re.search('set\(THIS (.*)\)', content).group(1)
-        self.name = name.strip().lower()
+        self.name = self.full_name.lower()
 
     def validate(self):
         if self.settings.compiler.get_safe("cppstd"):
@@ -52,30 +57,50 @@ class PackageConan(ConanFile):
 
     def requirements(self):
         # @+ START USER REQUIREMENTS
-        self.requires("information_model/[~0.4]@hahn-schickard/stable",
-                      headers=True, libs=True, transitive_headers=True,  transitive_libs=True)
-        self.requires('hasll/[~0.4]@hahn-schickard/stable', headers=True,
-                      libs=True, transitive_headers=True, transitive_libs=True)
+        self.requires("hasll/[~0.4]@hahn-schickard/stable",
+                      headers=True,
+                      libs=True,
+                      transitive_headers=True,
+                      transitive_libs=True
+                      )
+        self.requires("information_model/[~0.5]@hahn-schickard/stable",
+                      headers=True,
+                      libs=True,
+                      transitive_headers=True,
+                      transitive_libs=True
+                      )
+        self.requires("variant_visitor/[~0.2]@hahn-schickard/stable",
+                      visible=False
+                      )
         # @- END USER REQUIREMENTS
 
     def build_requirements(self):
-        self.test_requires("gtest/[~1.16]")
+        self.test_requires(
+            "information_model_mocks/[~0.1]@hahn-schickard/stable")
+        # @+ START USER BUILD REQUIREMENTS
+        # @- END USER BUILD REQUIREMENTS
 
     def configure(self):
         # @+ START USER REQUIREMENTS OPTION CONFIGURATION
-        self.options["gtest"].shared = True
+        pass
         # @- END USER REQUIREMENTS OPTION CONFIGURATION
 
     def layout(self):
         cmake_layout(self)
 
+    def config_options(self):
+        if self.settings.os == 'Windows':
+            del self.options.fPIC
+
     def generate(self):
         tc = CMakeToolchain(self)
+        tc.user_presets_path = False
         tc.variables['STATIC_CODE_ANALYSIS'] = False
         tc.variables['RUN_TESTS'] = False
         tc.variables['COVERAGE_TRACKING'] = False
         tc.variables['CMAKE_CONAN'] = False
-        tc.cache_variables["CMAKE_POLICY_DEFAULT_CMP0077"] = "NEW"
+        # @+ START USER CMAKE OPTIONS
+        # @- END USER CMAKE OPTIONS
         tc.generate()
 
     def build(self):
@@ -91,14 +116,14 @@ class PackageConan(ConanFile):
         copy(self, pattern='AUTHORS', dst='licenses', src=self.cwd)
 
     def package_info(self):
-        self.cpp_info.libs = []
+        self.cpp_info.libs = ["Data_Consumer_Adapter_Interface"]
         self.cpp_info.set_property("cmake_find_mode", "both")
         # @+ START USER DEFINES
-        project_name = to_camel_case(self.name)
+        self.cpp_info.requires = [
+            "hasll::hasll",
+            "information_model::information_model"
+        ]
         # @- END USER DEFINES
-        self.cpp_info.set_property("cmake_file_name", project_name)
-        cmake_target_name = project_name + "::" + project_name
+        self.cpp_info.set_property("cmake_file_name", self.full_name)
+        cmake_target_name = self.full_name + "::" + self.full_name
         self.cpp_info.set_property("cmake_target_name", cmake_target_name)
-
-    def package_id(self):
-        self.info.clear()